rename and generalize player module for multiple ships
